backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
minio==7.2.3
pymilvus==2.3.6
celery==5.3.6
redis==5.0.1
Pillow==10.2.0
opencv-python==4.9.0.80
ffmpeg-python==0.2.0
python-dotenv==1.0.0
torch>=2.0.0
torchvision>=0.15.0
cn-clip
